Add init.d/lxd.super script
It is very hacky, and I am still unhappy about how the stopall script
shuts down the daemon and then starts it up again, just to shut it
down cleanly.  However, supervise-daemon seems to be limited in the
ways it can shut down a daemon.

If only lxd had a signal spec...